requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Limiter==2.4.7
python-dotenv==0.19.0
email-validator==1.1.3
redis==4.5.1